+++ internal/parsevolume/parsevolume.go
@@ -1,4 +1,4 @@
-package loader
+package parsevolume
 
 import (
 	"strings"
