+++ internal/parsevolume/parsevolume_test.go
@@ -1,4 +1,4 @@
-package loader
+package parsevolume
 
 import (
 	"fmt"
